requirements.txt
openai>=1.0
python-dotenv>=1.0
rich>=13.0